requirements/third-party.txt
bcrypt==1.1.0
cffi==0.8.6
django-extensions==1.4.8
django-grappelli==2.6.3
django-phonenumber-field==0.6
djangorestframework==3.0.0
factory-boy==2.4.1
phonenumbers==7.0.1
protobuf==2.6.1
psycopg2==2.5.4
pycparser==2.10
schematics==1.0-0
six==1.8.0
wsgiref==0.1.2
git+ssh://git@github.com/django/django.git@e7b9a58b081299b30f807d5c66f7a5d1940efe4c
arrow==0.4.4
python-dateutil==2.4.0
pyotp==1.4.1
twilio==3.6.14
